requirements.txt
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==1.6.4
django-oscar-api==1.5.2
django-oscar-api-checkout==0.4.1
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.0
idna==2.7
mock==2.0.0
pbr==5.1.1
phonenumbers==8.10.0
phonenumberslite==8.10.0
Pillow==5.3.0
purl==1.4
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
